REFACTOR: NAV-47 - Rename impl package to router
