+++ src/main/java/ch/naviqore/raptor/router/Lookup.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 import java.util.Map;
