+++ src/main/java/ch/naviqore/raptor/router/RaptorLeg.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 import ch.naviqore.raptor.Leg;
 import lombok.AccessLevel;
