+++ src/main/java/ch/naviqore/raptor/router/RouteBuilder.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 import lombok.extern.log4j.Log4j2;
 import org.jetbrains.annotations.NotNull;
