+++ src/main/java/ch/naviqore/raptor/router/RouteStop.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 record RouteStop(int stopIndex, int routeIndex) {
 }
