+++ src/main/java/ch/naviqore/raptor/router/Stop.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 record Stop(String id, int stopRouteIdx, int numberOfRoutes, int sameStopTransferTime, int transferIdx,
             int numberOfTransfers) {
