+++ src/main/java/ch/naviqore/raptor/router/StopContext.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 record StopContext(Transfer[] transfers, Stop[] stops, int[] stopRoutes) {
 }
