+++ src/main/java/ch/naviqore/raptor/router/StopTime.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 record StopTime(int arrival, int departure) {
