+++ src/main/java/ch/naviqore/raptor/router/Transfer.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 record Transfer(int targetStopIdx, int duration) {
 }
