+++ src/test/java/ch/naviqore/raptor/router/RaptorRouterTest.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 import ch.naviqore.raptor.*;
 import org.junit.jupiter.api.BeforeEach;
