+++ src/test/java/ch/naviqore/raptor/router/RaptorRouterTestBuilder.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.impl;
+package ch.naviqore.raptor.router;
 
 import ch.naviqore.raptor.RaptorAlgorithm;
 import lombok.NoArgsConstructor;
